# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "01cloud-payment",
        "title": "01cloud-payment",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Payment Support",
            "email": "info@01cloud.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/payment/paymentsetting": {
            "get": {
                "description": "Get list of PaymentSetting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentSetting"
                ],
                "summary": "Get PaymentSetting",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/doc.PaymentSetting"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new PaymentSetting with the input paylod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentSetting"
                ],
                "summary": "Create a new PaymentSetting",
                "parameters": [
                    {
                        "description": "Create PaymentSetting",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/doc.PaymentSetting"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/doc.PaymentSetting"
                        }
                    }
                }
            }
        },
        "/payment/paymentsetting/{id}": {
            "get": {
                "description": "Get PaymentSetting by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentSetting"
                ],
                "summary": "Get PaymentSetting by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PaymentSetting id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/doc.PaymentSetting"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a PaymentSetting with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentSetting"
                ],
                "summary": "Update a PaymentSetting",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PaymentSetting id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update PaymentSetting",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/doc.PaymentSetting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/doc.PaymentSetting"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a PaymentSetting with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentSetting"
                ],
                "summary": "Delete a PaymentSetting",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PaymentSetting id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/doc.PaymentSetting"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "doc.PaymentSetting": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                },
                "promocode": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/doc.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "doc.User": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "company": {
                    "type": "string"
                },
                "designation": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "first_name": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}